+++ core/util/mask.ts
